Refactoring complete.  Now Extracting and Generating the final dataset is complete.   Now implementing the CsvWriter functionality

+++ LogFileGrowthTracker.Implementation/FileInfo/ServerLogFactInfo.cs
@@ -13,6 +13,6 @@ namespace ServerLogMonitorSystem.FileInfo
     {
         public uint FileId { get; set; }
         public DateTime TimeStamp { get; set; }
-        public uint SizeInBytes { get; set; }
+        public double SizeInBytes { get; set; }
     }
 }
